Alguns detalhes, como deixando o MuiTheme no App.js e outros detalhes
