+++ labeddit/src/Pages/LoginPage/index.js
@@ -13,19 +13,6 @@ import Footer from '../../components/Footer';
 import { useForm } from '../../hooks/useForm';
 import { UrlContext } from '../../contexts/UrlContext';
 import axios from 'axios';
-import { createMuiTheme, MuiThemeProvider } from "@material-ui/core";
-import Button from "@material-ui/core/Button";
-
-const MyTheme = createMuiTheme({
-  palette: {
-    primary: {
-      main: "#71EB1E",
-    },
-    secondary: {
-      main: "#ff9800",
-    },
-  },
-});
 
 const LoginPage = (props) => {
 
@@ -79,7 +66,6 @@ const LoginPage = (props) => {
 
   return (
     <LoginPageContainer>
-      <MuiThemeProvider theme={MyTheme}>
       <Header />
       <LoginFormContainer onSubmit={goToPrivateArea} >
         <LoginFormControl>
@@ -90,7 +76,6 @@ const LoginPage = (props) => {
             label={'E-mail'}
             onChange={handleInputChange}
             type='email'
-            
           />
         </LoginFormControl>
         <LoginFormControl>
@@ -101,28 +86,27 @@ const LoginPage = (props) => {
             label={'Senha'}
             onChange={handleInputChange}
             type='password'
-            
           />
         </LoginFormControl>
         <LoginButtonWrapper>
-          <Button 
+          <LoginButton 
             color="secondary"
             onClick={goToSignUp} 
-            >Cadastrar
-          </Button>
-          
+          >
+            Cadastrar
+          </LoginButton>
           <LoginButton 
             type='submit' 
             size='small'
             variant="contained"
             color="secondary"
-            >Acessar
+          >
+            Acessar
           </LoginButton>
           
         </LoginButtonWrapper>
       </LoginFormContainer>
       <Footer />
-      </MuiThemeProvider>
     </LoginPageContainer>
   )
 }
