+++ labeddit/src/components/Footer/style.js
@@ -2,11 +2,8 @@ import styled from 'styled-components';
 import {} from '@material-ui/core';
 
 export const FooterContainer = styled.footer`
-    
     height: 100px;
     border: 1px solid black;
-    position: absolute;
-    bottom: 0;
     width: 100%;
     color: black;
     background-color: #ff9800 ;
